changed product price field

+++ shop/models.py
@@ -4,5 +4,5 @@ from django.db import models
 class Product(models.Model):
     name = models.CharField(max_length=100)
     description = models.CharField(max_length=1000)
-    price = models.FloatField()
+    price = models.DecimalField(decimal_places=2, max_digits=11)
     stock = models.IntegerField()
